requirements.txt
Flask
pyopenssl
pymysql
requests
